worker/requirements.txt
pydantic>=2.7
psycopg[binary]>=3.2
redis>=5.0
rq>=1.16
httpx>=0.27
python-dotenv>=1.0
tenacity>=8.2
structlog>=24.1
orjson>=3.10
